requirements.txt
face_recognition==1.3.0
opencv-python==4.9.0.80
Flask==3.0.3
Pillow==10.3.0
pandas==2.2.2
numpy==1.26.4
